requirements.txt
quart==0.19.6
sqlalchemy==2.0.35
asyncpg==0.30.0
psutil==6.0.0
reportlab==4.2.2
openai==1.51.0
google-generativeai==0.8.3
tenacity==9.0.0
prometheus-client==0.21.0
stable-baselines3==2.3.2
numpy==1.26.4
scikit-learn==1.5.2
spacy==3.8.2
pybreaker==1.2.0
twilio==9.3.3
playwright==1.47.0
aiohttp==3.10.10
faker==30.8.2
deepgram-sdk==3.7.2
websockets==13.1
python-dotenv==1.0.1
aiokafka==0.8.0
beautifulsoup4==4.12.3
requests==2.32.3
hcp-vault-secrets